Reworking
New scripts using name space to attempt to keep things more organize. New animations are being implemented as well as the use of blending trees and a switch from movement by NavMesh to Character Controller.
Also a visual effect for a glow around the ball and a trial were implemented on this version together with a reorganization of folders and files withing the folder "3"
------:TODO:------
Add new animations and fix their time to fit the player actions.
Fix waiting times commanded coroutines that are not waiting.
Create a player game object to hold the starting parameters of the players and get implemented on "Home" screen/state.
Save player highest score locally in the system.