URP addition
Universal render pipeline added to the project in order to get better post-processing effects and VFXs.
A lot of changes to codes to make them easier to read. The function RoundTimer was extending too much so it was extracted to a new separate script to keep a constant structure on the world controller script.
Lights to indicate the edge of the platform, fireworks at the end of a round and a "Congrats!" text and animation was added on this version with the objective of creating a more immersive experience at the end of each round.

TODO
Make the game Harder (maybe by adding a down offset to the balls so they can chase the player on a different angle.)
Add sound effects and music to the game.
Change the skin and ball selection system so a integer can be saved as representation of the selected skin/ball.